Update the swagger config to more correctly implement the hosted docs API.
Why is this change needed?
--------------------------
Adding commands to automatically download hosted docs resulted in a weird multi-endpoint mess that blurred the line between hosted and async.

How does it address the issue?
------------------------------
This makes the final result of a hosted (sync or async) document a URL, and assumes that the user will use whatever normal download-from-url tools are in their environment to retrieve the document as desired.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------

+++ test/hosted_sync.rb
@@ -1,8 +1,14 @@
 require "bundler/setup"
 Bundler.require
+require "open-uri"
+
+api_key = File.read("#{__dir__}/../.docraptor_key").strip
+unless api_key
+  raise "Please put a valid (paid plan) api key in the .docraptor_key file when testing this feature."
+end
 
 DocRaptor.configure do |dr|
-  dr.username  = "YOUR_API_KEY_HERE"
+  dr.username  = api_key
   # dr.debugging = true
 end
 
@@ -11,15 +17,20 @@ $docraptor = DocRaptor::DocApi.new
 output_file = "hosted-ruby-sync.pdf"
 
 output_payload = $docraptor.create_hosted_doc(
-  test:             true,
+  test:             false,
   document_content: "<html><body>Hello from Ruby</body></html>",
   name:             output_file,
   document_type:    "pdf",
 )
 
-downloaded_document = $docraptor.get_doc(output_payload.download_id)
+# should be the unbranded hosted doc url
+unless output_payload.download_url.include?("documentdeliver")
+  raise "Returned URL was not the unbranded URL"
+end
 
-File.write(output_file, downloaded_document)
+actual_document = open output_payload.download_url
+IO.copy_stream(actual_document, output_file)
+
 output_type = `file -b #{output_file}`
 File.delete output_file
